Adds migration generator for CASinoCore implementations
Includes a fix for an improper engine mount location.

The `:as` option specifies the namespace for an Engine's named routes.
By default, this is the same as the Engine's `engine_name` property. In
the case of CASino, this would equal `casino_engine`, which is the
"down cased" string value of `CASino::Engine`

In order to keep this both simple and consistent, the `:as` attribute
should be in lowercase. Otherwise, the namespace would look like a
constant which would be inaccurate.

instalo generatore fu