Removes "API" acronym inflection
If CASino is installed into a Rails application that has its own Api
controllers, and it has not opted to add the inflected acronym of
"API", then CASino will break the host application.

In honor of the Principle of Least Surprise, CASino should not force
developers to change the behavior of their own application in order to
support the optional "API" acronym inflection.

In the case where the host application has its own acronym inflection,
this aliases `CASino::Api` and `CASino::Processor::Api` to `API` in
order to be as compatible as possible.